build: harden release flags, enable -Wextra, drop dead defines

+++ src/list_sort.c
@@ -139,7 +139,7 @@ void list_sort(void *priv, struct list_head *head,
 			part[lev] = NULL;
 		}
 		if (lev > max_lev) {
-			if (unlikely(lev >= ARRAY_SIZE(part)-1)) {
+			if (unlikely((size_t)lev >= ARRAY_SIZE(part)-1)) {
 				printk_once(KERN_DEBUG "list too long for efficiency\n");
 				lev--;
 			}
